Minor changes to solve issues founded on Scrutinizer - [X] Reviewed "Best Practice" 18 issues - [X] Reviewed "Bug" 145 issues - [X] Reviewed "Comprehensibility" 3 issues - [X] Reviewed "Documentation" 1 issues - [X] Reviewed "Security" 3 issues - [ ] Reviewed "Unused Code" 14 issues
Trying to reduce issues on 74 files with issues.

+++ Core/Model/AttachedFile.php
@@ -116,7 +116,9 @@ class AttachedFile extends Base\ModelClass
     public function delete()
     {
         if (parent::delete()) {
-            @\unlink(\FS_FOLDER . $this->path);
+            if (\file_exists(\FS_FOLDER . $this->path)) {
+                \unlink(\FS_FOLDER . $this->path);
+            }
             return true;
         }
